Хранение структур данных проекта вынесено в модуль structure.py, для общих функций создан модуль utils.py
